+++ src/scope/scope.js
@@ -638,7 +638,7 @@ export default class Scope {
     }));
     await this.objects.persist();
     const idsWithRemoteScope = componentIds.map(id => id.changeScope(remoteName));
-    return this.getMany(idsWithRemoteScope);
+    return this.getManyWithAllVersions(idsWithRemoteScope);
   }
 
   ensureDir() {
@@ -737,6 +737,7 @@ export default class Scope {
     const componentsObjects = await this.sources.getMany(componentsToUpdate);
     const componentsToUpdateP = componentsObjects.map(async (componentObjects) => {
       const component = componentObjects.component;
+      if (!component) return null;
       const latestVersion = await component.loadVersion(component.latest(), this.objects);
       let wasUpdated = false;
       latestVersion.dependencies.forEach((dependency) => {